Split ASM80 equ alias tests
